Support skipna in core.groupby_reduce
This is the right place to do it since skipna can be optionally True
for appropriate dtypes. It's hard to choose the right one when passing
Dataset to apply_ufunc. Instead we do it here where we are always
dealing with a single array and can choose based on dtype.