More efficient cohorts. (#165)
Closes #140

We apply the cohort "split" step after the blockwise reduction, then use
the tree reduction on each cohort.

We also use the `.blocks` accessor to index out blocks. This is still a
bit inefficient since we split by indexing out regular arrays, so we
could index out blocks that don't contain any cohort members. However,
because we are splitting after the blockwise reduction, the amount of
work duplication can be a lot less than splitting the bare array.

One side-effect is that "split-reduce" is now a synonym for "cohorts".
I see no benefit to having a separate code path.

We also sort the cohorts at detection time to minimize shuffling of the final
output when `sort=True` (the default). Finally, we avoid sorting if the groups 
are already sorted.